# contracts/tactical/v2/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tactical play",
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "/api/tactical/v2"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/campaigns/{cid}": {
      "get": {
        "operationId": "readTactical",
        "parameters": [
          {
            "name": "cid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "actor_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "TacticalSnapshotV2",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TacticalSnapshotV2"
                }
              }
            }
          },
          "400": {
            "description": "TacticalError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TacticalError"
                }
              }
            }
          },
          "401": {
            "description": "TacticalError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TacticalError"
                }
              }
            }
          },
          "403": {
            "description": "TacticalError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TacticalError"
                }
              }
            }
          },
          "404": {
            "description": "TacticalError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TacticalError"
                }
              }
            }
          },
          "409": {
            "description": "TacticalError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TacticalError"
                }
              }
            }
          },
          "413": {
            "description": "TacticalError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TacticalError"
                }
              }
            }
          },
          "429": {
            "description": "TacticalError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TacticalError"
                }
              }
            }
          }
        }
      }
    },
    "/campaigns/{cid}/commands": {
      "post": {
        "operationId": "executeTactical",
        "parameters": [
          {
            "name": "cid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TacticalRequestV2"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "TacticalSnapshotV2",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TacticalSnapshotV2"
                }
              }
            }
          },
          "400": {
            "description": "TacticalError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TacticalError"
                }
              }
            }
          },
          "401": {
            "description": "TacticalError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TacticalError"
                }
              }
            }
          },
          "403": {
            "description": "TacticalError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TacticalError"
                }
              }
            }
          },
          "404": {
            "description": "TacticalError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TacticalError"
                }
              }
            }
          },
          "409": {
            "description": "TacticalError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TacticalError"
                }
              }
            }
          },
          "413": {
            "description": "TacticalError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TacticalError"
                }
              }
            }
          },
          "429": {
            "description": "TacticalError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TacticalError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
...
      "TacticalSnapshotV2": {
        "additionalProperties": false,
        "properties": {
          "version": {
            "default": "tactical-v2",
            "title": "Version",
            "type": "string"
          },
          "campaign_id": {
            "title": "Campaign Id",
            "type": "string"
          },
          "actor_id": {
            "title": "Actor Id",
            "type": "string"
          },
          "revision": {
            "minimum": 0,
            "title": "Revision",
            "type": "integer"
          },
          "encounters": {
            "items": {
              "$ref": "#/components/schemas/TacticalEncounter"
            },
            "title": "Encounters",
            "type": "array"
          },
          "equipment": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/EquipmentView"
            },
            "title": "Equipment",
            "type": "array"
          }
        },
        "required": [
          "campaign_id",
          "actor_id",
          "revision",
          "encounters"
        ],
        "title": "TacticalSnapshotV2",
        "type": "object"
      },
...
      "ContinueCriticalMiss": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Id",
            "type": "string"
          },
          "actor_id": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Actor Id",
            "type": "string"
          },
          "expected_revision": {
            "minimum": 0,
            "title": "Expected Revision",
            "type": "integer"
          },
          "kind": {
            "const": "continue_critical_miss",
            "default": "continue_critical_miss",
            "title": "Kind",
            "type": "string"
          },
          "encounter_id": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Encounter Id",
            "type": "string"
          },
          "critical_id": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Critical Id",
            "type": "string"
          },
          "stage": {
            "enum": [
              "migrate",
              "resume"
            ],
            "title": "Stage",
            "type": "string"
          }
        },
        "required": [
          "id",
          "actor_id",
          "expected_revision",
          "encounter_id",
          "critical_id",
          "stage"
        ],
        "title": "ContinueCriticalMiss",
        "type": "object"
      },
...
      "TacticalRequestV2": {
        "additionalProperties": false,
        "properties": {
          "command": {
            "discriminator": {
              "mapping": {
                "choose_defense": "#/components/schemas/ChooseDefense",
                "continue_critical_miss": "#/components/schemas/ContinueCriticalMiss",
                "declare_thrown_landing": "#/components/schemas/DeclareThrownLanding",
                "migrate_encounter_hex": "#/components/schemas/MigrateEncounterHex",
                "repair_equipment": "#/components/schemas/RepairEquipment",
                "resolve_choke_effects": "#/components/schemas/ResolveChokeEffects",
                "resolve_weapon_explosion": "#/components/schemas/ResolveWeaponExplosion",
                "resume_interrupted_turn": "#/components/schemas/ResumeInterruptedTurn",
                "retrieve_equipment": "#/components/schemas/RetrieveEquipment",
                "take_combat_turn": "#/components/schemas/TakeCombatTurn",
                "take_unarmed_turn": "#/components/schemas/TakeUnarmedTurn"
              },
              "propertyName": "kind"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/TakeCombatTurn"
              },
              {
                "$ref": "#/components/schemas/TakeUnarmedTurn"
              },
              {
                "$ref": "#/components/schemas/ChooseDefense"
              },
              {
                "$ref": "#/components/schemas/MigrateEncounterHex"
              },
              {
                "$ref": "#/components/schemas/ResumeInterruptedTurn"
              },
              {
                "$ref": "#/components/schemas/ResolveChokeEffects"
              },
              {
                "$ref": "#/components/schemas/RepairEquipment"
              },
              {
                "$ref": "#/components/schemas/RetrieveEquipment"
              },
              {
                "$ref": "#/components/schemas/ContinueCriticalMiss"
              },
              {
                "$ref": "#/components/schemas/DeclareThrownLanding"
              },
              {
                "$ref": "#/components/schemas/ResolveWeaponExplosion"
              }
            ],
            "title": "Command"
          }
        },
        "required": [
          "command"
        ],
        "title": "TacticalRequestV2",
        "type": "object"
      },
      "TacticalError": {
        "type": "object",
        "required": [
          "code",
          "error"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "error": {
            "type": "string"
          }
        }
      }
    }
  }
}